requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.2
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.1
Markdown==3.3.6
psycopg2-binary==2.9.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.13.3
zipp==3.7.0
